third_party/blink/web_tests/fast/css/fontface-single-font-family-expected.txt

Test that @font-face has a single font-family-name.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.styleSheets[1].cssRules.length is 3
PASS document.styleSheets[1].cssRules[0].style.fontFamily is ""
PASS document.styleSheets[1].cssRules[1].style.fontFamily is "Ahem2"
PASS document.styleSheets[1].cssRules[2].style.fontFamily is ""
PASS successfullyParsed is true

TEST COMPLETE
